Move cat card sections from the HotspotForm to the HotspotDetails
This will get rid of any UX-related ambiguities. Before this, the HotspotForm contained the cat cards (which could be saved individually) and the hotspot save button (which could give the impression of atomicity of the hotspot changes).
